0623 pm1100

+++ Baedal/src/baedal/java/app/menus/Menu.java
@@ -9,7 +9,7 @@ import lombok.ToString;
 @ToString
 
 public class Menu {
-	private int storeNum;
+	private long storeNum;
 	private String menuName;
 	private int menuPrice;
 	private String menuContent;
